Keep plugins and load pre-load lookups via SPI
