+++ log4j-spring-cloud-config/log4j-spring-cloud-config-client/src/main/resources/META-INF/services/org.apache.logging.log4j.core.lookup.StrLookup
@@ -1,0 +1,1 @@
+org.apache.logging.log4j.spring.cloud.config.client.SpringLookup
