+++ log4j-web/src/main/resources/META-INF/services/org.apache.logging.log4j.core.lookup.StrLookup
@@ -1,0 +1,1 @@
+org.apache.logging.log4j.web.WebLookup
